+++ hutool-core/src/test/java/cn/hutool/core/date/DateUtilTest.java
@@ -1148,4 +1148,11 @@ public class DateUtilTest {
 		final Calendar c = DateUtil.calendar(date);
 		Assert.assertEquals(DateUtil.date(c), date);
 	}
+
+	@Test
+	public void issueI7H34NTest() {
+		final DateTime parse = DateUtil.parse("2019-10-22T09:56:03.000123Z");
+		Assert.assertNotNull(parse);
+		Assert.assertEquals("2019-10-22 09:56:03", parse.toString());
+	}
 }